# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: FassWerk API
  description: "API fuer Reservierung, Tischabrechnung, Lagerverwaltung, Bar-Administration\
    \ und Schichtabrechnung."
  contact:
    name: FassWerk Team
  version: 0.0.1-SNAPSHOT
servers:
- url: /
...
paths:
...
  /api/inventory/{id}:
    put:
      tags:
      - Inventory
      summary: Lagerartikel aktualisieren
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemRequest"
        required: true
      responses:
        "200":
          description: Lagerartikel aktualisiert
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InventoryItemResponse"
        "404":
          description: Lagerartikel nicht gefunden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InventoryItemResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Inventory
      summary: Lagerartikel loeschen
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: expectedRevision
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Lagerartikel geloescht
        "404":
          description: Lagerartikel nicht gefunden
      security:
      - bearerAuth: []
  /api/inventory/{id}/consumption-metadata:
    get:
      tags:
      - Inventory
      summary: Verbrauchsmetadaten fuer Lagerartikel abrufen
      operationId: getConsumptionMetadata
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Metadaten geladen oder leer bei fehlender Konfiguration
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConsumptionMetadataResponse"
      security:
      - bearerAuth: []
    put:
      tags:
      - Inventory
      summary: Verbrauchsmetadaten fuer Lagerartikel aktualisieren
      operationId: updateConsumptionMetadata
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsumptionMetadataRequest"
        required: true
      responses:
        "200":
          description: Verbrauchsmetadaten aktualisiert
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConsumptionMetadataResponse"
      security:
      - bearerAuth: []
  /api/inventory/configuration:
    get:
      tags:
      - Inventory
      summary: Globale Sales-/Business-Day-Konfiguration laden
      operationId: getConfiguration
      responses:
        "200":
          description: Konfiguration geladen
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SalesConfigurationResponse"
      security:
      - bearerAuth: []
    put:
      tags:
      - Inventory
      summary: Globale Sales-/Business-Day-Konfiguration speichern
      operationId: updateConfiguration
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesConfigurationRequest"
        required: true
      responses:
        "200":
          description: Konfiguration gespeichert
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SalesConfigurationResponse"
      security:
      - bearerAuth: []
...
  /api/inventory:
    get:
      tags:
      - Inventory
      summary: Lagerartikel auflisten
      operationId: list_2
      responses:
        "200":
          description: Lagerliste geladen
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItemResponse"
      security:
      - bearerAuth: []
    post:
      tags:
      - Inventory
      summary: Lagerartikel anlegen
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemRequest"
        required: true
      responses:
        "201":
          description: Lagerartikel erstellt
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InventoryItemResponse"
        "400":
          description: Ungueltige Eingabe
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InventoryItemResponse"
      security:
      - bearerAuth: []
  /api/inventory/{id}/calculate-reorder:
    post:
      tags:
      - Inventory
      summary: Nachbestellberechnung fuer Lagerartikel sofort ausfuehren
      operationId: calculateReorder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Berechnung ausgefuehrt
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReorderCalculationResponse"
      security:
      - bearerAuth: []
  /api/inventory/{id}/adjust:
    post:
      tags:
      - Inventory
      summary: Lagerbestand manuell korrigieren
      operationId: adjust
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryAdjustmentRequest"
        required: true
      responses:
        "200":
          description: Bestand angepasst
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InventoryItemResponse"
      security:
      - bearerAuth: []
  /api/inventory/configuration/manual-day-close:
    post:
      tags:
      - Inventory
      summary: Geschaeftstag manuell abschliessen
      operationId: closeBusinessDayManually
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualDayCloseRequest"
        required: true
      responses:
        "200":
          description: Geschaeftstag abgeschlossen
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SalesConfigurationResponse"
      security:
      - bearerAuth: []
...
  /api/reorder/orders/inventory/{inventoryItemId}:
    get:
      tags:
      - Reorder Orders
      summary: Nachbestellungen fuer Lagerartikel abrufen
      operationId: getReordersByInventoryItem
      parameters:
      - name: inventoryItemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Nachbestellungen geladen
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReorderOrderResponse"
      security:
      - bearerAuth: []
...
  /api/inventory/{id}/reorder-calculations/history:
    get:
      tags:
      - Inventory
      summary: Historie der Nachbestellberechnungen abrufen
      operationId: getReorderCalculationHistory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: Historie geladen
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReorderCalculationResponse"
      security:
      - bearerAuth: []
  /api/inventory/{id}/reorder-calculation:
    get:
      tags:
      - Inventory
      summary: Letzte Nachbestellberechnung fuer Lagerartikel abrufen
      operationId: getLatestReorderCalculation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Berechnung geladen oder null bei fehlenden Daten
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReorderCalculationResponse"
        "404":
          description: Lagerartikel nicht gefunden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReorderCalculationResponse"
      security:
      - bearerAuth: []
  /api/inventory/sales/weekly/{variantId}:
    get:
      tags:
      - Inventory
      summary: Woechentliche Verkaufsaggregation fuer Variante abrufen
      operationId: getWeeklySales
      parameters:
      - name: variantId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: weeks
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 4
      responses:
        "200":
          description: Woechentliche Verkaufsdaten geladen
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DrinkSalesWeeklyResponse"
      security:
      - bearerAuth: []
  /api/inventory/sales/daily:
    get:
      tags:
      - Inventory
      summary: Taegliche Verkaufsdaten fuer Zeitraum abrufen
      operationId: getDailySales
      parameters:
      - name: startDate
        in: query
        required: true
        schema:
          type: string
          format: date
      - name: endDate
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: Taegliche Verkaufsdaten geladen
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DrinkSalesDailyResponse"
      security:
      - bearerAuth: []
  /api/inventory/reorder-suggestions:
    get:
      tags:
      - Inventory
      summary: Nachbestellvorschlaege abrufen
      operationId: reorderSuggestions
      responses:
        "200":
          description: Nachbestellvorschlaege geladen
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReorderSuggestionResponse"
      security:
      - bearerAuth: []
  /api/inventory/reorder-calculations/below-threshold:
    get:
      tags:
      - Inventory
      summary: Artikel unter Schwellwert abrufen
      operationId: getItemsBelowThreshold
      responses:
        "200":
          description: Artikel unter Schwellwert geladen
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReorderCalculationResponse"
      security:
      - bearerAuth: []
  /api/inventory/movements:
    get:
      tags:
      - Inventory
      summary: Lagerbewegungen abrufen
      operationId: movements
      responses:
        "200":
          description: Bewegungen geladen
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryMovementResponse"
      security:
      - bearerAuth: []
  /api/inventory/defaults:
    get:
      tags:
      - Inventory
      summary: Standardwerte pro Gebindeart abrufen
      operationId: defaults
      responses:
        "200":
          description: Standardwerte geladen
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryPackageDefaultsResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ReorderOrderResponse:
      required:
      - createdAt
      - createdBy
      - id
      - inventoryItemId
      - inventoryItemName
      - notes
      - orderedQuantity
      - orderedUnit
      - receivedAt
      - receivedQuantity
      - scheduledDeliveryDate
      - scheduledDeliveryTime
      - status
      - supplierId
      - supplierName
      type: object
      properties:
        id:
          type: integer
          format: int64
        inventoryItemId:
          type: integer
          format: int64
        inventoryItemName:
          type: string
        supplierId:
          type: integer
          format: int64
        supplierName:
          type: string
        orderedQuantity:
          type: number
        orderedUnit:
          type: string
        scheduledDeliveryDate:
          type: string
          format: date
        scheduledDeliveryTime:
          type: string
          nullable: true
        status:
          type: string
        notes:
          type: string
          nullable: true
        receivedQuantity:
          type: number
          nullable: true
        receivedAt:
          type: string
          nullable: true
        createdBy:
          type: string
        createdAt:
          type: string
      description: Reorder Order Response contract
    InventoryItemRequest:
      required:
      - active
      - contentPerPackage
      - contentUnit
      - minimumStock
      - name
      - packageType
      - packagesInStock
      - reorderThreshold
      type: object
      properties:
        name:
          minLength: 1
          type: string
        linkedDrinkId:
          type: integer
          format: int64
          nullable: true
        linkedDrinkVariantId:
          type: integer
          format: int64
          nullable: true
        packageType:
          type: string
          enum:
          - BARREL
          - CRATE
          - BOTTLE
          - BOX
          - SINGLE_BOTTLE
        packagesInStock:
          minimum: 0.00
          exclusiveMinimum: false
          type: number
          description: Initial stock for creation; updates must retain the displayed
            package count. Use the adjustment endpoint for stock corrections.
        contentPerPackage:
          minimum: 0.01
          exclusiveMinimum: false
          type: number
        contentUnit:
          type: string
          enum:
          - LITER
          - MILLILITER
          - PIECE
        reorderThreshold:
          minimum: 0.00
          exclusiveMinimum: false
          type: number
        minimumStock:
          minimum: 0.00
          exclusiveMinimum: false
          type: number
        recommendedReorderAmount:
          minimum: 0.00
          exclusiveMinimum: false
          type: number
          nullable: true
        reorderThresholdPackages:
          minimum: 0.00
          exclusiveMinimum: false
          type: number
          nullable: true
        minimumStockPackages:
          minimum: 0.00
          exclusiveMinimum: false
          type: number
          nullable: true
        recommendedReorderPackages:
          minimum: 0.00
          exclusiveMinimum: false
          type: number
          nullable: true
        supplier:
          type: string
          nullable: true
        active:
          type: boolean
        expectedRevision:
          minimum: 0
          type: integer
          description: "Revision read before editing; required for updates, omitted\
            \ for creation"
          format: int64
          nullable: true
      description: Inventory Item Request contract
    InventoryItemResponse:
      required:
      - active
      - contentPerPackage
      - contentUnit
      - id
      - linkedDrinkId
      - linkedDrinkVariantId
      - minimumStock
      - name
      - packageType
      - packagesInStock
      - recommendedReorderAmount
      - reorderThreshold
      - revision
      - supplier
      - totalStockAmount
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        linkedDrinkId:
          type: integer
          format: int64
          nullable: true
        linkedDrinkVariantId:
          type: integer
          format: int64
          nullable: true
        packageType:
          type: string
          enum:
          - BARREL
          - CRATE
          - BOTTLE
          - BOX
          - SINGLE_BOTTLE
        packagesInStock:
          type: number
        contentPerPackage:
          type: number
        contentUnit:
          type: string
          enum:
          - LITER
          - MILLILITER
          - PIECE
        totalStockAmount:
          type: number
        reorderThreshold:
          type: number
        minimumStock:
          type: number
        recommendedReorderAmount:
          type: number
        supplier:
          type: string
          nullable: true
        active:
          type: boolean
        revision:
          type: integer
          format: int64
      description: Inventory Item Response contract
    ConsumptionMetadataRequest:
      type: object
      properties:
        leadTimeDays:
          minimum: 0
          type: integer
          format: int32
          nullable: true
        safetyStockFactor:
          minimum: 0.00
          exclusiveMinimum: false
          type: number
          nullable: true
        weeksLookback:
          minimum: 1
          type: integer
          format: int32
          nullable: true
      description: Consumption Metadata Request contract
    ConsumptionMetadataResponse:
      required:
      - id
      - inventoryItemId
      - leadTimeDays
      - safetyStockFactor
      - weeksLookback
      type: object
      properties:
        id:
          type: integer
          format: int64
        inventoryItemId:
          type: integer
          format: int64
        leadTimeDays:
          type: integer
          format: int32
        safetyStockFactor:
          type: number
        weeksLookback:
          type: integer
          format: int32
      description: Consumption Metadata Response contract
    SalesConfigurationRequest:
      type: object
      properties:
        weeksLookback:
          minimum: 1
          type: integer
          format: int32
          nullable: true
        defaultSafetyFactor:
          minimum: 0.00
          exclusiveMinimum: false
          type: number
          nullable: true
        defaultLeadTimeDays:
          minimum: 0
          type: integer
          format: int32
          nullable: true
        businessTimezone:
          type: string
          nullable: true
        businessDayEndsAt:
          type: string
          nullable: true
        manualBusinessDate:
          type: string
          format: date
          nullable: true
      description: Sales Configuration Request contract
    SalesConfigurationResponse:
      required:
      - businessDayEndsAt
      - businessTimezone
      - defaultLeadTimeDays
      - defaultSafetyFactor
      - effectiveBusinessDate
      - manualBusinessDate
      - weeksLookback
      type: object
      properties:
        weeksLookback:
          type: integer
          format: int32
        defaultSafetyFactor:
          type: number
        defaultLeadTimeDays:
          type: integer
          format: int32
        businessTimezone:
          type: string
        businessDayEndsAt:
          type: string
        manualBusinessDate:
          type: string
          format: date
          nullable: true
        effectiveBusinessDate:
          type: string
          format: date
      description: Sales Configuration Response contract
...
    ReorderCalculationResponse:
      required:
      - calculationDate
      - currentStockAmount
      - id
      - inventoryItemId
      - inventoryItemName
      - isBelowThreshold
      - recommendedReorderAmount
      - weeklyAverageConsumption
      - weeksUntilStockout
      type: object
      properties:
        id:
          type: integer
          format: int64
        inventoryItemId:
          type: integer
          format: int64
        inventoryItemName:
          type: string
        calculationDate:
          type: string
          format: date-time
        currentStockAmount:
          type: number
        weeklyAverageConsumption:
          type: number
        recommendedReorderAmount:
          type: number
        isBelowThreshold:
          type: boolean
        weeksUntilStockout:
          type: number
          nullable: true
      description: Reorder Calculation Response contract
    InventoryAdjustmentRequest:
      required:
      - amount
      - increase
      - reason
      type: object
      properties:
        amount:
          minimum: 0.01
          exclusiveMinimum: false
          type: number
        reason:
          minLength: 1
          type: string
        increase:
          type: boolean
      description: Inventory Adjustment Request contract
    ManualDayCloseRequest:
      required:
      - expectedBusinessDate
      type: object
      properties:
        expectedBusinessDate:
          type: string
          format: date
      description: Close the displayed business date exactly once for an idempotency
        key
...
    DrinkSalesWeeklyResponse:
      required:
      - averageDailyQuantity
      - averageDailyVolumeMl
      - drinkId
      - drinkName
      - drinkVariantId
      - drinkVariantName
      - id
      - quantitySold
      - volumeSoldMl
      - weekStartDate
      type: object
      properties:
        id:
          type: integer
          format: int64
        drinkId:
          type: integer
          format: int64
        drinkName:
          type: string
        drinkVariantId:
          type: integer
          format: int64
          nullable: true
        drinkVariantName:
          type: string
          nullable: true
        weekStartDate:
          type: string
          format: date
        quantitySold:
          type: number
        volumeSoldMl:
          type: number
        averageDailyQuantity:
          type: number
        averageDailyVolumeMl:
          type: number
      description: Drink Sales Weekly Response contract
    DrinkSalesDailyResponse:
      required:
      - drinkId
      - drinkName
      - drinkVariantId
      - drinkVariantName
      - id
      - quantitySold
      - saleDate
      - volumeSoldMl
      type: object
      properties:
        id:
          type: integer
          format: int64
        drinkId:
          type: integer
          format: int64
        drinkName:
          type: string
        drinkVariantId:
          type: integer
          format: int64
          nullable: true
        drinkVariantName:
          type: string
          nullable: true
        saleDate:
          type: string
          format: date
        quantitySold:
          type: number
        volumeSoldMl:
          type: number
      description: Drink Sales Daily Response contract
    ReorderSuggestionResponse:
      required:
      - contentUnit
      - currentStock
      - inventoryItemId
      - inventoryItemName
      - minimumStock
      - packageSize
      - packageType
      - recommendedOrderAmount
      - recommendedOrderPackages
      - supplier
      - threshold
      type: object
      properties:
        inventoryItemId:
          type: integer
          format: int64
        inventoryItemName:
          type: string
        currentStock:
          type: number
        threshold:
          type: number
        minimumStock:
          type: number
        recommendedOrderAmount:
          type: number
        recommendedOrderPackages:
          type: number
        packageSize:
          type: number
        packageType:
          type: string
          enum:
          - BARREL
          - CRATE
          - BOTTLE
          - BOX
          - SINGLE_BOTTLE
        contentUnit:
          type: string
          enum:
          - LITER
          - MILLILITER
          - PIECE
        supplier:
          type: string
          nullable: true
      description: Reorder Suggestion Response contract
...
  securitySchemes:
    bearerAuth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: JWT